+++ openslp/libslp/libslp_dereg.c
@@ -211,7 +211,7 @@ SLPError SLPDereg(SLPHandle  hSLP,
     /* check for invalid parameters */
     /*------------------------------*/
     if( hSLP        == 0 ||
-        *(unsigned long*)hSLP != SLP_HANDLE_SIG ||
+        *(unsigned int*)hSLP != SLP_HANDLE_SIG ||
         srvUrl      == 0 ||
         *srvUrl     == 0 ||  /* url can't be empty string */
         callback    == 0)
